+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - Use `expm1` in `n_thermal` for stability. ([#717], [#718])
-- Bump compat for `CUDA.jl` to v6. ([#687])
+- Use `CUDACore` and `cuSPARSE` as dependencies instead of `CUDA`. ([#687])
 
 ## [v0.46.0]
 Release date: 2026-05-11